Week6/requirements.txt
Flask
PyJWT
Authlib
requests
python-dotenv
